assets: move to pkg directory
So we can use the asset server from other projects.
