+++ internal/cli/wire.go
@@ -13,7 +13,6 @@ import (
 	"k8s.io/client-go/tools/clientcmd/api"
 
 	"github.com/windmilleng/tilt/internal/analytics"
-	"github.com/windmilleng/tilt/internal/assets"
 	"github.com/windmilleng/tilt/internal/build"
 	"github.com/windmilleng/tilt/internal/container"
 	"github.com/windmilleng/tilt/internal/demo"
@@ -29,6 +28,7 @@ import (
 	"github.com/windmilleng/tilt/internal/store"
 	tft "github.com/windmilleng/tilt/internal/tft/client"
 	"github.com/windmilleng/tilt/internal/tiltfile"
+	"github.com/windmilleng/tilt/pkg/assets"
 )
 
 var K8sWireSet = wire.NewSet(
