+++ internal/hud/server/server.go
@@ -13,13 +13,13 @@ import (
 	"github.com/windmilleng/wmclient/pkg/analytics"
 
 	tiltanalytics "github.com/windmilleng/tilt/internal/analytics"
-	"github.com/windmilleng/tilt/internal/assets"
 	"github.com/windmilleng/tilt/internal/hud/webview"
 	"github.com/windmilleng/tilt/internal/logger"
 	"github.com/windmilleng/tilt/internal/model"
 	"github.com/windmilleng/tilt/internal/sail/client"
 	"github.com/windmilleng/tilt/internal/store"
 	tft "github.com/windmilleng/tilt/internal/tft/client"
+	"github.com/windmilleng/tilt/pkg/assets"
 )
 
 const tiltAlertsDomain = "alerts.tilt.dev"
