+++ internal/hud/server/server_test.go
@@ -13,12 +13,12 @@ import (
 	"github.com/windmilleng/wmclient/pkg/analytics"
 
 	tiltanalytics "github.com/windmilleng/tilt/internal/analytics"
-	"github.com/windmilleng/tilt/internal/assets"
 	"github.com/windmilleng/tilt/internal/hud/server"
 	"github.com/windmilleng/tilt/internal/model"
 	"github.com/windmilleng/tilt/internal/sail/client"
 	"github.com/windmilleng/tilt/internal/store"
 	tft "github.com/windmilleng/tilt/internal/tft/client"
+	"github.com/windmilleng/tilt/pkg/assets"
 )
 
 func TestHandleAnalyticsEmptyRequest(t *testing.T) {
