+++ pkg/assets/server.go
